requirements.txt
# EXIF/GPS reading
exifread>=3.0.0

# Image processing
Pillow>=10.0.0


